test/my_project/requirements.txt
asgiref==3.3.4
Cerberus==1.3.4
Django==3.2.4
djangorestframework==3.12.4
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
